Update build.gradle (#70)
* Update build.gradle

Linking the Javadoc to the Java 6 API seems to create a strange bug with certain Java JDKs
Do we really need it?

* Update build.gradle

It seems the bug is gone when linking to a Java greater Java 10. Since we are currently building with Java 12 or higher, I think we can link to the Java 12 JavaDoc.